Move VM execute test to PublicVirtualMachineSuite
It wasn't obvious to me that the `PublicByteCodeOperationSuite.scala`
test belonged in the `bc` package, but I found the test valuable from
the point of view that it tests whether the VM executes an instruction
successfully.

Accordingly, add a new test to `PublicVirtualMachineSuite.scala` to test
the `execute()` method on `VirtualMachineConcrete` that runs a simple
script adding two integers together.