Make ByteCode classes always return new VM
The ByteCode classes were returning the original VM, rather than a new
one as mentioned in the comments. Doing so also seems more idiomatic in
Scala, since we want to avoid mutable state where possible.

Changes `var` declarations to `val` to convert the mutable variables
into immutable values.

Every time we perform an action on the VM, we now use the most
recently-returned VM to perform the next action.

I also added a space in between the arithmetic operators which removes
the need to do `res1*(-1)`; instead we can simplify it to `res1 * -1`.